fix(review): run artifact-violation guard on all execution paths
